Hedron is now a static librarry (.lib) instead of a dynamic library (.dll)
Cleaned up the premake files
Cleaned up the project so now there's no warnings when compiling

+++ Hedron/src/Hedron/ImGui/ImGuiLayer.cpp
@@ -81,7 +81,7 @@ namespace Hedron
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::get();
-		io.DisplaySize = ImVec2(app.get_window().get_width(), app.get_window().get_height());
+		io.DisplaySize = ImVec2((float)app.get_window().get_width(), (float)app.get_window().get_height());
 
 		// Rendering
 		ImGui::Render();
